Compiles on ESP8266 & ESP32
ESP compilers don't like mixed argument types to min() and max() functions. This fixes that. New Adafruit_ImageReader release requires this in order to compile, even if only using SD card.

+++ src/Adafruit_SPIFlash.cpp
@@ -238,7 +238,7 @@ uint32_t Adafruit_SPIFlash::writeBuffer (uint32_t address, uint8_t const *buffer
 	  waitUntilReady();
 	  writeEnable();
 
-	  uint16_t const toWrite = min(remain, SFLASH_PAGE_SIZE);
+	  uint16_t const toWrite = min(remain, (uint32_t)SFLASH_PAGE_SIZE);
 
 		if ( !_trans->writeMemory(address, buffer, toWrite) ) break;
